pyproject.toml
[tool.numpydoc_validation]
checks = ["all"]

[tool.black]
line-length = 79
target-version = ['py310', 'py311', 'py312', 'py313']

[tool.check-manifest]
ignore = [
    "tox.ini",
    ".header-template",
    "requirements.txt",
    "requirements_dev.txt",
    "tests",
    "assets",
    "draft",
    "tools",
    "res",
]

[tool.pytest.ini_options]
markers = [
    "plot: marks tests related to matplotlib integration (deselect with '-m \"not plot\"')",
]
testpaths = ["tests"]
addopts = "-m 'not slow' -n auto -p no:warnings"

# =============================================================================
# PACKAGING
# =============================================================================

[project]
name = "garpar"
version = "1.5.0dev0"
authors = [
    { name = "Diego N Gimenez Irusta", email = "d.gimenez0101@unc.edu.ar" },
    { name = "Nadia A Luczywo", email = "nadia.luczywo@unc.edu.ar" },
    { name = "Juan B Cabral", email = "jbcabral@unc.edu.ar" }
]
readme = "README.md"
license = { file = "LICENSE.txt" }
description = "Generation and Analysis of Real and Artificial Portfolio Returns"
keywords = [
    "market simulation",
    "informational efficiency",
    "portfolio optimization",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Scientific/Engineering/Finances",
]
urls = { Homepage = "https://garpar.quatrope.org/", Repository = "https://github.com/quatrope/garpar" }
dependencies = [
    "attrs>=23.2.0",
    "build>=1.2.1",
    "h5py>=3.11.0",
    "importlib_metadata",
    "numpy>=2.2.0",
    "pyportfolioopt>=1.5.0",
    "scikit-learn>=1.6.0",
    "seaborn>=0.13.0",
    "pandas>=2.2.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["garpar", "garpar.*"]
namespaces = false
